Nwm czemu ale przestało działać na prodzie ale naprawiłem i działa xdddd

+++ Absencja.cs
@@ -112,7 +112,7 @@ namespace Excel_Data_Importer_WARS
                         {
                             using (SqlCommand command = new(DbManager.Update_Dzien_Pracy_Strefa, DbManager.GetConnection(), DbManager.Transaction_Manager.CurrentTransaction))
                             {
-                                command.Parameters.Add("@NowaStrefa", SqlDbType.Int).Value = Helper.Strefa.Czas_Pracy_W_Delegacji;
+                                command.Parameters.Add("@NowaStrefa", SqlDbType.Int).Value = (int)Helper.Strefa.Czas_Pracy_W_Delegacji;
                                 command.Parameters.Add("@IdDniaGodz", SqlDbType.Int).Value = Dzien_Godz;
                                 command.Parameters.Add("@NewOdGodz", SqlDbType.DateTime).Value = DbManager.Base_Date + TimeSpan.FromHours(8);
                                 command.Parameters.Add("@NewDoGodz", SqlDbType.DateTime).Value = DbManager.Base_Date + TimeSpan.FromHours(8) + TimeSpan.FromHours((double)absencja.Liczba_Godzin_Przepracowanych);
