+++ Reader_Karta_Ewidencji_Konduktora_v1.cs
@@ -537,7 +537,7 @@ namespace Excel_Data_Importer_WARS
                     command.Parameters.Add("@GodzDoDate", SqlDbType.DateTime).Value = godzDoDate;
                     command.Parameters.Add("@DataInsert", SqlDbType.DateTime).Value = Data_Karty;
                     command.Parameters.Add("@PRI_PraId", SqlDbType.Int).Value = IdPracownika;
-                    command.Parameters.Add("@Strefa", SqlDbType.Int).Value = Strefa;
+                    command.Parameters.Add("@Strefa", SqlDbType.Int).Value = (int)Strefa;
                     duplicate = (int)command.ExecuteScalar() == 1;
                 }
 
@@ -547,7 +547,7 @@ namespace Excel_Data_Importer_WARS
                     command.Parameters.Add("@GodzDoDate", SqlDbType.DateTime).Value = godzDoDate;
                     command.Parameters.Add("@DataInsert", SqlDbType.DateTime).Value = Data_Karty;
                     command.Parameters.Add("@PRI_PraId", SqlDbType.Int).Value = IdPracownika;
-                    command.Parameters.Add("@Strefa", SqlDbType.Int).Value = Helper.Strefa.Czas_Pracy_W_Delegacji;
+                    command.Parameters.Add("@Strefa", SqlDbType.Int).Value = (int)Helper.Strefa.Czas_Pracy_W_Delegacji;
                     duplicateDE = (int)command.ExecuteScalar() == 1;
                 }
 
@@ -559,7 +559,7 @@ namespace Excel_Data_Importer_WARS
                         command.Parameters.Add("@GodzDoDate", SqlDbType.DateTime).Value = godzDoDate;
                         command.Parameters.Add("@DataInsert", SqlDbType.DateTime).Value = Data_Karty;
                         command.Parameters.Add("@PRI_PraId", SqlDbType.Int).Value = IdPracownika;
-                        command.Parameters.Add("@Strefa", SqlDbType.Int).Value = Strefa;
+                        command.Parameters.Add("@Strefa", SqlDbType.Int).Value = (int)Strefa;
                         command.Parameters.Add("@ImieMod", SqlDbType.NVarChar, 20).Value = Helper.Truncate(Internal_Error_Logger.Last_Mod_Osoba, 20);
                         command.Parameters.Add("@NazwiskoMod", SqlDbType.NVarChar, 50).Value = Helper.Truncate(Internal_Error_Logger.Last_Mod_Osoba, 50);
                         command.Parameters.Add("@DataMod", SqlDbType.DateTime).Value = Internal_Error_Logger.Last_Mod_Time;
